+++ src/accessibility.cpp
@@ -96,10 +96,27 @@ Accessibility::precomputeRangeQueries(float radius) {
 }
 
 
-std::vector<int>
+vector<int>
 Accessibility::Route(int src, int tgt, int graphno) {
     vector<NodeID> ret = this->ga[graphno]->Route(src, tgt);
     return vector<int> (ret.begin(), ret.end());
+}
+
+
+vector<vector<int>>
+Accessibility::Routes(vector<long> sources, vector<long> targets, int graphno) {
+
+    int n = std::min(sources.size(), targets.size()); // in case lists don't match
+    vector<vector<int>> routes(n);
+
+    #pragma omp parallel
+    #pragma omp for schedule(guided)
+    for (int i = 0 ; i < n ; i++) {
+        vector<NodeID> ret = this->ga[graphno]->Route(sources[i], targets[i], 
+            omp_get_thread_num());
+        routes[i] = vector<int> (ret.begin(), ret.end());
+    }
+    return routes;
 }
 
 
@@ -109,11 +126,11 @@ Accessibility::Distance(int src, int tgt, int graphno) {
 }
 
 
-std::vector<double>
+vector<double>
 Accessibility::Distances(vector<long> sources, vector<long> targets, int graphno) {                       
     
     int n = std::min(sources.size(), targets.size()); // in case lists don't match
-    vector<double> distances (n);
+    vector<double> distances(n);
     
     #pragma omp parallel
     #pragma omp for schedule(guided)